WH-A1, Resetup Project into Maven
